requirements.txt
Pillow
numpy
onnxruntime